@@ -3,7 +3,7 @@ name = "mls-crypto-provider"
 description = "MLS Crypto Provider wrapping core-crypto-keystore"
 repository = "https://github.com/wireapp/core-crypto"
 version = "0.11.0"
-edition = "2018"
+edition = "2021"
 license = "GPL-3.0-only"
 publish = false
 
@@ -16,12 +16,14 @@ default = []
 raw-rand-access = [] # TESTING ONLY
 
 [dependencies]
-openmls_traits = { version = "0.1", features = ["single-threaded"] }
+openmls_traits = "0.1"
+tls_codec = { version = "0.3.0-pre.3", features = ["derive", "serde", "mls"] }
 aes-gcm = "0.10"
 sha2 = "0.10"
 chacha20poly1305 = "0.10"
 hmac = "0.12"
-ed25519-dalek = "1.0"
+ed25519-dalek = "2.0.0-rc.2"
+signature = "2.1"
 ecdsa = { version = "0.16", features = ["der"] }
 p256 = "0.13"
 p384 = "0.13"
@@ -30,12 +32,14 @@ p384 = "0.13"
 hkdf = "0.12"
 rand = { version = "0.8", features = ["getrandom"] }
 getrandom = { version = "0.2", features = ["js"] }
+rand_core = "0.6"
 rand_chacha = "0.3"
-hpke = { version = "0.2", package = "hpke-rs", default-features = false, features = ["hazmat", "serialization"] }
-hpke-rs-crypto = "0.2"
-hpke-rs-rust-crypto = "0.2"
 zeroize = "1.5"
 thiserror = "1.0"
+
+[dependencies.hpke]
+version = "0.10"
+features = ["x25519", "p256", "p384", "serde_impls"]
 
 [target.'cfg(not(target_os = "ios"))'.dependencies]
 core-crypto-keystore = { version = "^0.11.0", path = "../keystore" }
@@ -46,7 +50,7 @@ core-crypto-keystore = { version = "^0.11.0", path = "../keystore", features = [
 [dev-dependencies]
 wasm-bindgen-test = "0.3"
 uuid = { version = "1.0", features = ["v4", "js"] }
-openmls = { version = "0.4", default-features = false }
+openmls = { version = "0.20", default-features = false }
 rstest = "0.17"
 rstest_reuse = "0.5"
 async-std = { version = "1.12", features = ["attributes"] }
