@@ -1,48 +1,16 @@
-use crate::prelude::{CryptoResult, MlsCredentialType, MlsError, WireIdentity};
-use openmls::prelude::{Credential, CredentialBundle};
-use tls_codec::Serialize;
-use wire_e2e_identity::prelude::WireIdentityReader;
+use crate::prelude::{CryptoError, CryptoResult, MlsCredentialType};
+use openmls::prelude::{Credential, CredentialType};
 
 pub(crate) trait CredentialExt {
-    fn keystore_key(&self) -> CryptoResult<Vec<u8>>;
-    fn get_type(&self) -> MlsCredentialType;
-    fn extract_identity(&self) -> Option<WireIdentity>;
-}
-
-impl CredentialExt for CredentialBundle {
-    #[inline(always)]
-    fn keystore_key(&self) -> CryptoResult<Vec<u8>> {
-        self.credential().keystore_key()
-    }
-
-    fn get_type(&self) -> MlsCredentialType {
-        self.credential().get_type()
-    }
-
-    fn extract_identity(&self) -> Option<WireIdentity> {
-        self.credential().extract_identity()
-    }
+    fn get_type(&self) -> CryptoResult<MlsCredentialType>;
 }
 
 impl CredentialExt for Credential {
-    fn keystore_key(&self) -> CryptoResult<Vec<u8>> {
-        Ok(self.signature_key().tls_serialize_detached().map_err(MlsError::from)?)
-    }
-
-    fn get_type(&self) -> MlsCredentialType {
-        match self.credential {
-            openmls::prelude::MlsCredentialType::Basic(_) => MlsCredentialType::Basic,
-            openmls::prelude::MlsCredentialType::X509(_) => MlsCredentialType::X509,
-        }
-    }
-
-    fn extract_identity(&self) -> Option<WireIdentity> {
-        match &self.credential {
-            openmls::prelude::MlsCredentialType::X509(c) => {
-                let leaf = c.cert_chain.get(0)?;
-                Some(leaf.as_slice().extract_identity().ok()?.into())
-            }
-            _ => None,
+    fn get_type(&self) -> CryptoResult<MlsCredentialType> {
+        match self.credential_type() {
+            CredentialType::Basic => Ok(MlsCredentialType::Basic),
+            CredentialType::X509 => Ok(MlsCredentialType::X509),
+            CredentialType::Unknown(_) => Err(CryptoError::ImplementationError),
         }
     }
 }
