@@ -14,9 +14,11 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see http://www.gnu.org/licenses/.
 
+use openmls_traits::OpenMlsCryptoProvider;
 use std::collections::HashMap;
 
-use openmls::prelude::KeyPackage;
+use mls_crypto_provider::MlsCryptoProvider;
+use openmls::prelude::{KeyPackage, KeyPackageIn};
 use tls_codec::Deserialize;
 
 use crate::{
@@ -41,9 +43,11 @@ impl ConversationMember {
     ///
     /// # Errors
     /// Deserialization errors
-    pub fn new_raw(client_id: ClientId, kp_ser: Vec<u8>) -> CryptoResult<Self> {
-        use openmls::prelude::TlsDeserializeTrait as _;
-        let kp = KeyPackage::tls_deserialize(&mut &kp_ser[..]).map_err(MlsError::from)?;
+    pub fn new_raw(client_id: ClientId, kp_ser: Vec<u8>, backend: &MlsCryptoProvider) -> CryptoResult<Self> {
+        let kp = KeyPackageIn::tls_deserialize_bytes(kp_ser).map_err(MlsError::from)?;
+        let kp = kp
+            .validate(backend.crypto(), openmls::versions::ProtocolVersion::Mls10)
+            .map_err(MlsError::from)?;
 
         Ok(Self {
             id: client_id.to_vec(),
@@ -87,9 +91,12 @@ impl ConversationMember {
     ///
     /// # Errors
     /// Deserialization errors
-    pub fn add_keypackage(&mut self, kp: Vec<u8>) -> CryptoResult<()> {
-        let kp = KeyPackage::tls_deserialize(&mut &kp[..]).map_err(MlsError::from)?;
-        let cid = ClientId::from(kp.credential().identity());
+    pub fn add_keypackage(&mut self, kp: Vec<u8>, backend: &MlsCryptoProvider) -> CryptoResult<()> {
+        let kp = KeyPackageIn::tls_deserialize_bytes(kp).map_err(MlsError::from)?;
+        let kp = kp
+            .validate(backend.crypto(), openmls::versions::ProtocolVersion::Mls10)
+            .map_err(MlsError::from)?;
+        let cid = ClientId::from(kp.leaf_node().credential().identity());
         self.clients.entry(cid).or_insert_with(Vec::new).push(kp);
         Ok(())
     }
@@ -128,7 +135,7 @@ impl ConversationMember {
     pub fn random_generate_clientless(
         case: &crate::test_utils::TestCase,
         backend: &mls_crypto_provider::MlsCryptoProvider,
-    ) -> CryptoResult<(Self, openmls::prelude::CredentialBundle)> {
+    ) -> CryptoResult<(Self, crate::mls::credential::CredentialBundle)> {
         let (credential, client_id) = match case.credential_type {
             crate::prelude::MlsCredentialType::Basic => {
                 let user_uuid = uuid::Uuid::new_v4();
@@ -166,13 +173,15 @@ impl ConversationMember {
 
 #[cfg(test)]
 pub mod tests {
-    use openmls::prelude::*;
     use wasm_bindgen_test::wasm_bindgen_test;
 
     use mls_crypto_provider::MlsCryptoProvider;
 
-    use crate::prelude::key_package::INITIAL_KEYING_MATERIAL_COUNT;
-    use crate::{mls::ClientId, test_utils::*};
+    use crate::{
+        prelude::{key_package::INITIAL_KEYING_MATERIAL_COUNT, ClientId},
+        test_utils::*,
+    };
+    use openmls::prelude::{CryptoConfig, KeyPackage};
 
     use super::ConversationMember;
 
@@ -206,18 +215,20 @@ pub mod tests {
         #[wasm_bindgen_test]
         pub async fn add_valid_keypackage_should_add_it_to_client(case: TestCase) {
             let backend = MlsCryptoProvider::try_new_in_memory("test").await.unwrap();
-            let (mut member, credential) = ConversationMember::random_generate_clientless(&case, &backend).unwrap();
+            let (mut member, cb) = ConversationMember::random_generate_clientless(&case, &backend).unwrap();
             let cid = ClientId::from(member.id.as_slice());
 
-            let (kp, _) = KeyPackageBundle::new(&[case.ciphersuite().0], &credential, &backend, vec![])
-                .unwrap()
-                .into_parts();
+            let cfg = CryptoConfig::with_default_version(case.ciphersuite().into());
+            let kp = KeyPackage::builder()
+                .build(cfg, &backend, &cb.signature_key.clone(), cb.into())
+                .await
+                .unwrap();
 
             let mut kp_raw = vec![];
             use tls_codec::Serialize as _;
             kp.tls_serialize(&mut kp_raw).unwrap();
             assert!(member.clients.get(&cid).is_none());
-            assert!(member.add_keypackage(kp_raw).is_ok());
+            assert!(member.add_keypackage(kp_raw, &backend).is_ok());
             assert!(member.clients.get(&cid).is_some());
         }
 
@@ -227,7 +238,7 @@ pub mod tests {
             let backend = MlsCryptoProvider::try_new_in_memory("test").await.unwrap();
             let (mut member, _) = ConversationMember::random_generate_clientless(&case, &backend).unwrap();
             let previous_clients = member.clients.clone();
-            assert!(member.add_keypackage(b"invalid-keypackage".to_vec()).is_err());
+            assert!(member.add_keypackage(b"invalid-keypackage".to_vec(), &backend).is_err());
             // ensure clients are not altered in the process
             assert_eq!(member.clients, previous_clients);
         }
