@@ -16,62 +16,62 @@
 
 use crate::{
     connection::{DatabaseConnection, KeystoreDatabaseConnection},
-    entities::{Entity, EntityBase, EntityFindParams, ProteusPrekey, StringEntityId},
+    entities::{Entity, EntityBase, EntityFindParams, MlsPskBundle, StringEntityId},
     CryptoKeystoreResult, MissingKeyErrorKind,
 };
 
 #[async_trait::async_trait(?Send)]
-impl EntityBase for ProteusPrekey {
+impl EntityBase for MlsPskBundle {
     type ConnectionType = KeystoreDatabaseConnection;
 
     fn to_missing_key_err_kind() -> MissingKeyErrorKind {
-        MissingKeyErrorKind::ProteusPrekey
+        MissingKeyErrorKind::MlsPskBundle
     }
 
     async fn find_all(conn: &mut Self::ConnectionType, params: EntityFindParams) -> CryptoKeystoreResult<Vec<Self>> {
         let storage = conn.storage();
-        storage.get_all("proteus_prekeys", Some(params)).await
+        storage.get_all("mls_psk_bundles", Some(params)).await
     }
 
     async fn save(&self, conn: &mut Self::ConnectionType) -> crate::CryptoKeystoreResult<()> {
         let storage = conn.storage_mut();
-        storage.save("proteus_prekeys", &mut [self.clone()]).await
+        storage.save("mls_psk_bundles", &mut [self.clone()]).await?;
+
+        Ok(())
     }
 
     async fn find_one(
         conn: &mut Self::ConnectionType,
         id: &StringEntityId,
     ) -> crate::CryptoKeystoreResult<Option<Self>> {
-        let storage = conn.storage();
-        storage.get("proteus_prekeys", id.as_slice()).await
+        conn.storage().get("mls_psk_bundles", id.as_slice()).await
     }
 
     async fn count(conn: &mut Self::ConnectionType) -> crate::CryptoKeystoreResult<usize> {
-        let storage = conn.storage();
-        storage.count("proteus_prekeys").await
+        conn.storage().count("mls_psk_bundles").await
     }
 
     async fn delete(conn: &mut Self::ConnectionType, ids: &[StringEntityId]) -> crate::CryptoKeystoreResult<()> {
         let storage = conn.storage_mut();
         let ids: Vec<Vec<u8>> = ids.iter().map(StringEntityId::to_bytes).collect();
-        storage.delete("proteus_prekeys", &ids).await
+        storage.delete("mls_psk_bundles", &ids).await
     }
 }
 
-impl Entity for ProteusPrekey {
+impl Entity for MlsPskBundle {
     fn id_raw(&self) -> &[u8] {
-        self.id_bytes()
+        self.psk_id.as_slice()
     }
 
     fn encrypt(&mut self, cipher: &aes_gcm::Aes256Gcm) -> CryptoKeystoreResult<()> {
-        self.prekey = Self::encrypt_data(cipher, self.prekey.as_slice(), self.aad())?;
-        Self::ConnectionType::check_buffer_size(self.prekey.len())?;
+        self.psk = Self::encrypt_data(cipher, self.psk.as_slice(), self.aad())?;
+        Self::ConnectionType::check_buffer_size(self.psk.len())?;
 
         Ok(())
     }
 
     fn decrypt(&mut self, cipher: &aes_gcm::Aes256Gcm) -> CryptoKeystoreResult<()> {
-        self.prekey = Self::decrypt_data(cipher, self.prekey.as_slice(), self.aad())?;
+        self.psk = Self::decrypt_data(cipher, self.psk.as_slice(), self.aad())?;
 
         Ok(())
     }
