@@ -1,3 +1,6 @@
+use crate::EntropySeed;
+use crate::MlsProviderError;
+use rand_core::{RngCore, SeedableRng};
 use std::sync::RwLock;
 
 use aes_gcm::{
@@ -5,24 +8,17 @@ use aes_gcm::{
     Aes128Gcm, Aes256Gcm, KeyInit,
 };
 use chacha20poly1305::ChaCha20Poly1305;
-use ecdsa::{signature::Verifier, Signature, SigningKey, VerifyingKey};
-use ed25519_dalek::Signer as DalekSigner;
 use hkdf::Hkdf;
-use hpke::Hpke;
-use hpke_rs_crypto::types as hpke_types;
-use hpke_rs_rust_crypto::HpkeRustCrypto;
 use openmls_traits::{
     crypto::OpenMlsCrypto,
     random::OpenMlsRand,
     types::{
-        self, AeadType, Ciphersuite, CryptoError, HashType, HpkeAeadType, HpkeCiphertext, HpkeConfig, HpkeKdfType,
-        HpkeKemType, HpkeKeyPair, SignatureScheme,
+        self, AeadType, Ciphersuite, CryptoError, ExporterSecret, HashType, HpkeAeadType, HpkeConfig, HpkeKdfType,
+        HpkeKemType, SignatureScheme,
     },
 };
-use rand::{RngCore, SeedableRng};
 use sha2::{Digest, Sha256, Sha384, Sha512};
-
-use crate::{EntropySeed, MlsProviderError};
+use tls_codec::SecretVLBytes;
 
 #[derive(Debug)]
 pub struct RustCrypto {
@@ -32,7 +28,7 @@ pub struct RustCrypto {
 impl Default for RustCrypto {
     fn default() -> Self {
         Self {
-            rng: rand_chacha::ChaCha20Rng::from_entropy().into(),
+            rng: RwLock::new(rand_chacha::ChaCha20Rng::from_entropy()),
         }
     }
 }
@@ -42,63 +38,6 @@ impl RustCrypto {
         Self {
             rng: rand_chacha::ChaCha20Rng::from_seed(seed.0).into(),
         }
-    }
-}
-
-#[cfg(feature = "raw-rand-access")]
-impl RustCrypto {
-    pub fn borrow_rand(&self) -> std::sync::RwLockWriteGuard<rand_chacha::ChaCha20Rng> {
-        self.rng.write().unwrap()
-    }
-}
-
-impl OpenMlsRand for RustCrypto {
-    type Error = MlsProviderError;
-
-    fn random_array<const N: usize>(&self) -> Result<[u8; N], Self::Error> {
-        let mut rng = self.rng.write().map_err(|_| MlsProviderError::RngLockPoison)?;
-        let mut out = [0u8; N];
-        rng.try_fill_bytes(&mut out)
-            .map_err(|_| MlsProviderError::UnsufficientEntropy)?;
-        Ok(out)
-    }
-
-    fn random_vec(&self, len: usize) -> Result<Vec<u8>, Self::Error> {
-        let mut rng = self.rng.write().map_err(|_| MlsProviderError::RngLockPoison)?;
-        let mut out = vec![0u8; len];
-        rng.try_fill_bytes(&mut out)
-            .map_err(|_| MlsProviderError::UnsufficientEntropy)?;
-        Ok(out)
-    }
-}
-
-#[inline]
-fn kem_mode(kem: HpkeKemType) -> hpke_types::KemAlgorithm {
-    match kem {
-        HpkeKemType::DhKemP256 => hpke_types::KemAlgorithm::DhKemP256,
-        HpkeKemType::DhKemP384 => hpke_types::KemAlgorithm::DhKemP384,
-        HpkeKemType::DhKemP521 => hpke_types::KemAlgorithm::DhKemP521,
-        HpkeKemType::DhKem25519 => hpke_types::KemAlgorithm::DhKem25519,
-        HpkeKemType::DhKem448 => hpke_types::KemAlgorithm::DhKem448,
-    }
-}
-
-#[inline]
-fn kdf_mode(kdf: HpkeKdfType) -> hpke_types::KdfAlgorithm {
-    match kdf {
-        HpkeKdfType::HkdfSha256 => hpke_types::KdfAlgorithm::HkdfSha256,
-        HpkeKdfType::HkdfSha384 => hpke_types::KdfAlgorithm::HkdfSha384,
-        HpkeKdfType::HkdfSha512 => hpke_types::KdfAlgorithm::HkdfSha512,
-    }
-}
-
-#[inline]
-fn aead_mode(aead: HpkeAeadType) -> hpke_types::AeadAlgorithm {
-    match aead {
-        HpkeAeadType::AesGcm128 => hpke_types::AeadAlgorithm::Aes128Gcm,
-        HpkeAeadType::AesGcm256 => hpke_types::AeadAlgorithm::Aes256Gcm,
-        HpkeAeadType::ChaCha20Poly1305 => hpke_types::AeadAlgorithm::ChaCha20Poly1305,
-        HpkeAeadType::Export => hpke_types::AeadAlgorithm::HpkeExport,
     }
 }
 
@@ -127,7 +66,7 @@ impl OpenMlsCrypto for RustCrypto {
         hash_type: openmls_traits::types::HashType,
         salt: &[u8],
         ikm: &[u8],
-    ) -> Result<Vec<u8>, openmls_traits::types::CryptoError> {
+    ) -> Result<SecretVLBytes, openmls_traits::types::CryptoError> {
         match hash_type {
             HashType::Sha2_256 => Ok(Hkdf::<Sha256>::extract(Some(salt), ikm).0.as_slice().into()),
             HashType::Sha2_384 => Ok(Hkdf::<Sha384>::extract(Some(salt), ikm).0.as_slice().into()),
@@ -141,27 +80,36 @@ impl OpenMlsCrypto for RustCrypto {
         prk: &[u8],
         info: &[u8],
         okm_len: usize,
-    ) -> Result<Vec<u8>, openmls_traits::types::CryptoError> {
-        let mut okm = vec![0u8; okm_len];
+    ) -> Result<SecretVLBytes, openmls_traits::types::CryptoError> {
         match hash_type {
             HashType::Sha2_256 => {
                 let hkdf = Hkdf::<Sha256>::from_prk(prk).map_err(|_| CryptoError::HkdfOutputLengthInvalid)?;
+
+                let mut okm = vec![0u8; okm_len];
                 hkdf.expand(info, &mut okm)
                     .map_err(|_| CryptoError::HkdfOutputLengthInvalid)?;
+
+                Ok(okm.into())
             }
             HashType::Sha2_512 => {
                 let hkdf = Hkdf::<Sha512>::from_prk(prk).map_err(|_| CryptoError::HkdfOutputLengthInvalid)?;
+
+                let mut okm = vec![0u8; okm_len];
                 hkdf.expand(info, &mut okm)
                     .map_err(|_| CryptoError::HkdfOutputLengthInvalid)?;
+
+                Ok(okm.into())
             }
             HashType::Sha2_384 => {
                 let hkdf = Hkdf::<Sha384>::from_prk(prk).map_err(|_| CryptoError::HkdfOutputLengthInvalid)?;
+
+                let mut okm = vec![0u8; okm_len];
                 hkdf.expand(info, &mut okm)
                     .map_err(|_| CryptoError::HkdfOutputLengthInvalid)?;
+
+                Ok(okm.into())
             }
         }
-
-        Ok(okm)
     }
 
     fn hash(
@@ -186,23 +134,26 @@ impl OpenMlsCrypto for RustCrypto {
     ) -> Result<Vec<u8>, openmls_traits::types::CryptoError> {
         match alg {
             AeadType::Aes128Gcm => {
-                let aes = Aes128Gcm::new_from_slice(key).map_err(|_| CryptoError::InvalidLength)?;
+                let aes = Aes128Gcm::new_from_slice(key).map_err(|_| CryptoError::CryptoLibraryError)?;
+
                 aes.encrypt(nonce.into(), Payload { msg: data, aad })
                     .map(|r| r.as_slice().into())
-                    .map_err(|_| CryptoError::AeadEncryptionError)
+                    .map_err(|_| CryptoError::CryptoLibraryError)
             }
             AeadType::Aes256Gcm => {
-                let aes = Aes256Gcm::new_from_slice(key).map_err(|_| CryptoError::InvalidLength)?;
+                let aes = Aes256Gcm::new_from_slice(key).map_err(|_| CryptoError::CryptoLibraryError)?;
+
                 aes.encrypt(nonce.into(), Payload { msg: data, aad })
                     .map(|r| r.as_slice().into())
-                    .map_err(|_| CryptoError::AeadEncryptionError)
+                    .map_err(|_| CryptoError::CryptoLibraryError)
             }
             AeadType::ChaCha20Poly1305 => {
-                let chacha_poly = ChaCha20Poly1305::new_from_slice(key).map_err(|_| CryptoError::InvalidLength)?;
+                let chacha_poly = ChaCha20Poly1305::new_from_slice(key).map_err(|_| CryptoError::CryptoLibraryError)?;
+
                 chacha_poly
                     .encrypt(nonce.into(), Payload { msg: data, aad })
                     .map(|r| r.as_slice().into())
-                    .map_err(|_| CryptoError::AeadEncryptionError)
+                    .map_err(|_| CryptoError::CryptoLibraryError)
             }
         }
     }
@@ -217,19 +168,19 @@ impl OpenMlsCrypto for RustCrypto {
     ) -> Result<Vec<u8>, openmls_traits::types::CryptoError> {
         match alg {
             AeadType::Aes128Gcm => {
-                let aes = Aes128Gcm::new_from_slice(key).map_err(|_| CryptoError::InvalidLength)?;
+                let aes = Aes128Gcm::new_from_slice(key).map_err(|_| CryptoError::CryptoLibraryError)?;
                 aes.decrypt(nonce.into(), Payload { msg: ct_tag, aad })
                     .map(|r| r.as_slice().into())
                     .map_err(|_| CryptoError::AeadDecryptionError)
             }
             AeadType::Aes256Gcm => {
-                let aes = Aes256Gcm::new_from_slice(key).map_err(|_| CryptoError::InvalidLength)?;
+                let aes = Aes256Gcm::new_from_slice(key).map_err(|_| CryptoError::CryptoLibraryError)?;
                 aes.decrypt(nonce.into(), Payload { msg: ct_tag, aad })
                     .map(|r| r.as_slice().into())
                     .map_err(|_| CryptoError::AeadDecryptionError)
             }
             AeadType::ChaCha20Poly1305 => {
-                let chacha_poly = ChaCha20Poly1305::new_from_slice(key).map_err(|_| CryptoError::InvalidLength)?;
+                let chacha_poly = ChaCha20Poly1305::new_from_slice(key).map_err(|_| CryptoError::CryptoLibraryError)?;
                 chacha_poly
                     .decrypt(nonce.into(), Payload { msg: ct_tag, aad })
                     .map(|r| r.as_slice().into())
@@ -242,59 +193,25 @@ impl OpenMlsCrypto for RustCrypto {
         &self,
         alg: openmls_traits::types::SignatureScheme,
     ) -> Result<(Vec<u8>, Vec<u8>), openmls_traits::types::CryptoError> {
+        let mut rng = self.rng.write().map_err(|_| CryptoError::InsufficientRandomness)?;
+
         match alg {
             SignatureScheme::ECDSA_SECP256R1_SHA256 => {
-                let mut rng = self.rng.write().map_err(|_| CryptoError::InsufficientRandomness)?;
-                let k = p256::ecdsa::SigningKey::random(&mut *rng);
-                let pk = k.verifying_key().to_encoded_point(false).as_bytes().into();
-                Ok((k.to_bytes().as_slice().into(), pk))
+                let sk = p256::ecdsa::SigningKey::random(&mut *rng);
+                let pk = sk.verifying_key().to_encoded_point(false).to_bytes().into();
+                Ok((sk.to_bytes().to_vec(), pk))
             }
             SignatureScheme::ECDSA_SECP384R1_SHA384 => {
-                let mut rng = self.rng.write().map_err(|_| CryptoError::InsufficientRandomness)?;
-                let k = p384::ecdsa::SigningKey::random(&mut *rng);
-                let pk = k.verifying_key().to_encoded_point(false).as_bytes().into();
-                Ok((k.to_bytes().as_slice().into(), pk))
-            }
-            SignatureScheme::ECDSA_SECP521R1_SHA512 => {
-                // TODO: Uncomment this once p521 crate is ready
-                // let mut rng = self.rng.write().map_err(|_| CryptoError::InsufficientRandomness)?;
-                // let k = p521::ecdsa::SigningKey::random(&mut *rng);
-                // let pk = k.verifying_key().to_encoded_point(false).as_bytes().into();
-                // Ok((k.to_bytes().as_slice().into(), pk))
-                Err(CryptoError::UnsupportedSignatureScheme)
+                let sk = p384::ecdsa::SigningKey::random(&mut *rng);
+                let pk = sk.verifying_key().to_encoded_point(false).to_bytes().into();
+                Ok((sk.to_bytes().to_vec(), pk))
             }
             SignatureScheme::ED25519 => {
-                let mut rng = self.rng.write().map_err(|_| CryptoError::InsufficientRandomness)?;
-                // TODO: Once ed25519_dalek updates its `rand` dependency to rand_core 0.6, uncomment the following line and get rid of this mess
-                // let kp = ed25519_dalek::Keypair::generate(&mut rng);
-                let mut sk_bytes = [0u8; ed25519_dalek::SECRET_KEY_LENGTH];
-                rng.fill_bytes(&mut sk_bytes);
-                let sk = ed25519_dalek::SecretKey::from_bytes(sk_bytes.as_slice())
-                    .map_err(|_| CryptoError::InvalidLength)?;
-                let pk: ed25519_dalek::PublicKey = (&sk).into();
-                let mut kp_bytes = [0u8; ed25519_dalek::KEYPAIR_LENGTH];
-
-                let sk_bytes = sk.to_bytes();
-                if sk_bytes.len() != ed25519_dalek::SECRET_KEY_LENGTH {
-                    return Err(CryptoError::InvalidLength);
-                }
-                kp_bytes[..ed25519_dalek::SECRET_KEY_LENGTH].copy_from_slice(&sk_bytes);
-
-                let pk_bytes = pk.to_bytes();
-                if pk_bytes.len() != ed25519_dalek::KEYPAIR_LENGTH - ed25519_dalek::SECRET_KEY_LENGTH {
-                    return Err(CryptoError::InvalidLength);
-                }
-                kp_bytes[ed25519_dalek::SECRET_KEY_LENGTH..ed25519_dalek::KEYPAIR_LENGTH].copy_from_slice(&pk_bytes);
-
-                let k = ed25519_dalek::Keypair::from_bytes(kp_bytes.as_slice())
-                    .map_err(|_| CryptoError::InvalidLength)?
-                    .to_bytes();
-                let pk = k[ed25519_dalek::SECRET_KEY_LENGTH..].to_vec();
-                // full key here because we need it to sign...
-                let sk_pk = k.into();
-                Ok((sk_pk, pk))
+                let k = ed25519_dalek::SigningKey::generate(&mut *rng);
+                let pk = k.verifying_key();
+                Ok((k.to_bytes().into(), pk.to_bytes().into()))
             }
-            SignatureScheme::ED448 => Err(CryptoError::UnsupportedSignatureScheme),
+            _ => Err(CryptoError::UnsupportedSignatureScheme),
         }
     }
 
@@ -305,59 +222,32 @@ impl OpenMlsCrypto for RustCrypto {
         pk: &[u8],
         signature: &[u8],
     ) -> Result<(), openmls_traits::types::CryptoError> {
+        use signature::Verifier as _;
         match alg {
             SignatureScheme::ECDSA_SECP256R1_SHA256 => {
-                let k: p256::ecdsa::VerifyingKey = VerifyingKey::from_encoded_point(
-                    &p256::EncodedPoint::from_bytes(pk).map_err(|_| CryptoError::InvalidLength)?,
-                )
-                .map_err(|_| CryptoError::InvalidLength)?;
+                let k = p256::ecdsa::VerifyingKey::from_sec1_bytes(pk).map_err(|_| CryptoError::CryptoLibraryError)?;
 
-                k.verify(
-                    data,
-                    &Signature::from_der(signature).map_err(|_| CryptoError::InvalidSignature)?,
-                )
-                .map_err(|_| CryptoError::InvalidSignature)
+                let signature =
+                    p256::ecdsa::DerSignature::from_bytes(signature).map_err(|_| CryptoError::InvalidSignature)?;
+
+                k.verify(data, &signature).map_err(|_| CryptoError::InvalidSignature)
             }
             SignatureScheme::ECDSA_SECP384R1_SHA384 => {
-                let k: p384::ecdsa::VerifyingKey = VerifyingKey::from_encoded_point(
-                    &p384::EncodedPoint::from_bytes(pk).map_err(|_| CryptoError::InvalidLength)?,
-                )
-                .map_err(|_| CryptoError::InvalidLength)?;
+                let k = p384::ecdsa::VerifyingKey::from_sec1_bytes(pk).map_err(|_| CryptoError::CryptoLibraryError)?;
 
-                k.verify(
-                    data,
-                    &Signature::from_der(signature).map_err(|_| CryptoError::InvalidSignature)?,
-                )
-                .map_err(|_| CryptoError::InvalidSignature)
-            }
-            SignatureScheme::ECDSA_SECP521R1_SHA512 => {
-                // TODO: Uncomment this once p521 crate is ready
-                // let k: p521::ecdsa::VerifyingKey = VerifyingKey::from_encoded_point(
-                //     &p521::EncodedPoint::from_bytes(pk).map_err(|_| CryptoError::InvalidLength)?,
-                // )
-                // .map_err(|_| CryptoError::InvalidLength)?;
+                let signature =
+                    p384::ecdsa::DerSignature::from_bytes(signature).map_err(|_| CryptoError::InvalidSignature)?;
 
-                // k.verify(
-                //     data,
-                //     &Signature::from_der(signature).map_err(|_| CryptoError::InvalidSignature)?,
-                // )
-                // .map_err(|_| CryptoError::InvalidSignature)
-                Err(CryptoError::UnsupportedSignatureScheme)
+                k.verify(data, &signature).map_err(|_| CryptoError::InvalidSignature)
             }
             SignatureScheme::ED25519 => {
-                let k = ed25519_dalek::PublicKey::from_bytes(pk).map_err(|_| CryptoError::InvalidLength)?;
-                if signature.len() != ed25519_dalek::SIGNATURE_LENGTH {
-                    return Err(CryptoError::InvalidLength);
-                }
-                let mut sig = [0u8; ed25519_dalek::SIGNATURE_LENGTH];
-                sig.clone_from_slice(signature);
-                k.verify_strict(
-                    data,
-                    &ed25519_dalek::Signature::from_bytes(&sig).map_err(|_| CryptoError::InvalidLength)?,
-                )
-                .map_err(|_| CryptoError::InvalidSignature)
+                let k = ed25519_dalek::VerifyingKey::try_from(pk).map_err(|_| CryptoError::CryptoLibraryError)?;
+
+                let sig = ed25519_dalek::Signature::from_slice(signature).map_err(|_| CryptoError::InvalidSignature)?;
+
+                k.verify_strict(data, &sig).map_err(|_| CryptoError::InvalidSignature)
             }
-            SignatureScheme::ED448 => Err(CryptoError::UnsupportedSignatureScheme),
+            _ => Err(CryptoError::UnsupportedSignatureScheme),
         }
     }
 
@@ -367,27 +257,41 @@ impl OpenMlsCrypto for RustCrypto {
         data: &[u8],
         key: &[u8],
     ) -> Result<Vec<u8>, openmls_traits::types::CryptoError> {
-        use ecdsa::signature::Signer as _;
+        use signature::Signer as _;
+
         match alg {
             SignatureScheme::ECDSA_SECP256R1_SHA256 => {
-                let k: p256::ecdsa::SigningKey =
-                    SigningKey::from_bytes(key.into()).map_err(|_| CryptoError::InvalidLength)?;
-                let signature: p256::ecdsa::Signature = k.sign(data);
-                Ok(signature.to_der().to_bytes().into())
-            }
-            SignatureScheme::ECDSA_SECP384R1_SHA384 => {
-                let k: p384::ecdsa::SigningKey =
-                    SigningKey::from_bytes(key.into()).map_err(|_| CryptoError::InvalidLength)?;
-                let signature: p384::ecdsa::Signature = k.sign(data);
-                Ok(signature.to_der().to_bytes().into())
-            }
-            SignatureScheme::ECDSA_SECP521R1_SHA512 => Err(CryptoError::UnsupportedSignatureScheme),
-            SignatureScheme::ED25519 => {
-                let k = ed25519_dalek::Keypair::from_bytes(key).map_err(|_| CryptoError::InvalidLength)?;
-                let signature = k.sign(data);
+                let k = p256::ecdsa::SigningKey::from_bytes(key.into()).map_err(|_| CryptoError::CryptoLibraryError)?;
+                let signature: p256::ecdsa::DerSignature =
+                    k.try_sign(data).map_err(|_| CryptoError::CryptoLibraryError)?;
                 Ok(signature.to_bytes().into())
             }
-            SignatureScheme::ED448 => Err(CryptoError::UnsupportedSignatureScheme),
+            SignatureScheme::ECDSA_SECP384R1_SHA384 => {
+                let k = p384::ecdsa::SigningKey::from_bytes(key.into()).map_err(|_| CryptoError::CryptoLibraryError)?;
+                let signature: p384::ecdsa::DerSignature =
+                    k.try_sign(data).map_err(|_| CryptoError::CryptoLibraryError)?;
+                Ok(signature.to_bytes().into())
+            }
+            SignatureScheme::ED25519 => {
+                let k = match key.len() {
+                    // Compat layer for legacy keypairs [seed, pk]
+                    ed25519_dalek::KEYPAIR_LENGTH => {
+                        let mut sk = zeroize::Zeroizing::new([0u8; ed25519_dalek::KEYPAIR_LENGTH]);
+                        sk.copy_from_slice(key);
+                        ed25519_dalek::SigningKey::from_keypair_bytes(&sk)
+                            .map_err(|_| CryptoError::CryptoLibraryError)?
+                    }
+                    ed25519_dalek::SECRET_KEY_LENGTH => {
+                        let mut sk = zeroize::Zeroizing::new([0u8; ed25519_dalek::SECRET_KEY_LENGTH]);
+                        sk.copy_from_slice(key);
+                        ed25519_dalek::SigningKey::from_bytes(&sk)
+                    }
+                    _ => return Err(CryptoError::CryptoLibraryError),
+                };
+                let signature = k.try_sign(data).map_err(|_| CryptoError::CryptoLibraryError)?;
+                Ok(signature.to_bytes().into())
+            }
+            _ => Err(CryptoError::UnsupportedSignatureScheme),
         }
     }
 
@@ -398,13 +302,31 @@ impl OpenMlsCrypto for RustCrypto {
         info: &[u8],
         aad: &[u8],
         ptxt: &[u8],
-    ) -> types::HpkeCiphertext {
-        let (kem_output, ciphertext) = hpke_from_config(config)
-            .seal(&pk_r.into(), info, aad, ptxt, None, None, None)
-            .unwrap();
-        HpkeCiphertext {
-            kem_output: kem_output.into(),
-            ciphertext: ciphertext.into(),
+    ) -> Result<types::HpkeCiphertext, CryptoError> {
+        let mut rng = self.rng.write().map_err(|_| CryptoError::InsufficientRandomness)?;
+
+        match config {
+            HpkeConfig(HpkeKemType::DhKem25519, HpkeKdfType::HkdfSha256, HpkeAeadType::AesGcm128) => {
+                hpke_core::hpke_seal::<hpke::aead::AesGcm128, hpke::kdf::HkdfSha256, hpke::kem::X25519HkdfSha256>(
+                    pk_r, info, aad, ptxt, &mut *rng,
+                )
+            }
+            HpkeConfig(HpkeKemType::DhKem25519, HpkeKdfType::HkdfSha256, HpkeAeadType::ChaCha20Poly1305) => {
+                hpke_core::hpke_seal::<hpke::aead::ChaCha20Poly1305, hpke::kdf::HkdfSha256, hpke::kem::X25519HkdfSha256>(
+                    pk_r, info, aad, ptxt, &mut *rng,
+                )
+            }
+            HpkeConfig(HpkeKemType::DhKemP256, HpkeKdfType::HkdfSha256, HpkeAeadType::AesGcm128) => {
+                hpke_core::hpke_seal::<hpke::aead::AesGcm128, hpke::kdf::HkdfSha256, hpke::kem::DhP256HkdfSha256>(
+                    pk_r, info, aad, ptxt, &mut *rng,
+                )
+            }
+            HpkeConfig(HpkeKemType::DhKemP384, HpkeKdfType::HkdfSha384, HpkeAeadType::AesGcm256) => {
+                hpke_core::hpke_seal::<hpke::aead::AesGcm256, hpke::kdf::HkdfSha384, hpke::kem::DhP384HkdfSha384>(
+                    pk_r, info, aad, ptxt, &mut *rng,
+                )
+            }
+            _ => Err(CryptoError::UnsupportedKem),
         }
     }
 
@@ -416,18 +338,47 @@ impl OpenMlsCrypto for RustCrypto {
         info: &[u8],
         aad: &[u8],
     ) -> Result<Vec<u8>, CryptoError> {
-        hpke_from_config(config)
-            .open(
-                input.kem_output.as_slice(),
-                &sk_r.into(),
-                info,
-                aad,
-                input.ciphertext.as_slice(),
-                None,
-                None,
-                None,
-            )
-            .map_err(|_| CryptoError::HpkeDecryptionError)
+        let plaintext = match config {
+            HpkeConfig(HpkeKemType::DhKem25519, HpkeKdfType::HkdfSha256, HpkeAeadType::AesGcm128) => {
+                hpke_core::hpke_open::<hpke::aead::AesGcm128, hpke::kdf::HkdfSha256, hpke::kem::X25519HkdfSha256>(
+                    sk_r,
+                    input.kem_output.as_slice(),
+                    info,
+                    aad,
+                    input.ciphertext.as_slice(),
+                )?
+            }
+            HpkeConfig(HpkeKemType::DhKem25519, HpkeKdfType::HkdfSha256, HpkeAeadType::ChaCha20Poly1305) => {
+                hpke_core::hpke_open::<hpke::aead::ChaCha20Poly1305, hpke::kdf::HkdfSha256, hpke::kem::X25519HkdfSha256>(
+                    sk_r,
+                    input.kem_output.as_slice(),
+                    info,
+                    aad,
+                    input.ciphertext.as_slice(),
+                )?
+            }
+            HpkeConfig(HpkeKemType::DhKemP256, HpkeKdfType::HkdfSha256, HpkeAeadType::AesGcm128) => {
+                hpke_core::hpke_open::<hpke::aead::AesGcm128, hpke::kdf::HkdfSha256, hpke::kem::DhP256HkdfSha256>(
+                    sk_r,
+                    input.kem_output.as_slice(),
+                    info,
+                    aad,
+                    input.ciphertext.as_slice(),
+                )?
+            }
+            HpkeConfig(HpkeKemType::DhKemP384, HpkeKdfType::HkdfSha384, HpkeAeadType::AesGcm256) => {
+                hpke_core::hpke_open::<hpke::aead::AesGcm256, hpke::kdf::HkdfSha384, hpke::kem::DhP384HkdfSha384>(
+                    sk_r,
+                    input.kem_output.as_slice(),
+                    info,
+                    aad,
+                    input.ciphertext.as_slice(),
+                )?
+            }
+            _ => return Err(CryptoError::UnsupportedKem),
+        };
+
+        Ok(plaintext)
     }
 
     fn hpke_setup_sender_and_export(
@@ -437,14 +388,45 @@ impl OpenMlsCrypto for RustCrypto {
         info: &[u8],
         exporter_context: &[u8],
         exporter_length: usize,
-    ) -> Result<(Vec<u8>, Vec<u8>), CryptoError> {
-        let (kem_output, context) = hpke_from_config(config)
-            .setup_sender(&pk_r.into(), info, None, None, None)
-            .map_err(|_| CryptoError::SenderSetupError)?;
-        let exported_secret = context
-            .export(exporter_context, exporter_length)
-            .map_err(|_| CryptoError::ExporterError)?;
-        Ok((kem_output, exported_secret))
+    ) -> Result<(Vec<u8>, ExporterSecret), openmls_traits::types::CryptoError> {
+        let mut rng = self.rng.write().map_err(|_| CryptoError::InsufficientRandomness)?;
+
+        let (kem_output, export) =
+            match config {
+                HpkeConfig(HpkeKemType::DhKem25519, HpkeKdfType::HkdfSha256, HpkeAeadType::AesGcm128) => {
+                    hpke_core::hpke_export_tx::<
+                        hpke::aead::AesGcm128,
+                        hpke::kdf::HkdfSha256,
+                        hpke::kem::X25519HkdfSha256,
+                    >(pk_r, info, exporter_context, exporter_length, &mut *rng)?
+                }
+                HpkeConfig(HpkeKemType::DhKem25519, HpkeKdfType::HkdfSha256, HpkeAeadType::ChaCha20Poly1305) => {
+                    hpke_core::hpke_export_tx::<
+                        hpke::aead::ChaCha20Poly1305,
+                        hpke::kdf::HkdfSha256,
+                        hpke::kem::X25519HkdfSha256,
+                    >(pk_r, info, exporter_context, exporter_length, &mut *rng)?
+                }
+                HpkeConfig(HpkeKemType::DhKemP256, HpkeKdfType::HkdfSha256, HpkeAeadType::AesGcm128) => {
+                    hpke_core::hpke_export_tx::<
+                        hpke::aead::AesGcm128,
+                        hpke::kdf::HkdfSha256,
+                        hpke::kem::DhP256HkdfSha256,
+                    >(pk_r, info, exporter_context, exporter_length, &mut *rng)?
+                }
+                HpkeConfig(HpkeKemType::DhKemP384, HpkeKdfType::HkdfSha384, HpkeAeadType::AesGcm256) => {
+                    hpke_core::hpke_export_tx::<
+                        hpke::aead::AesGcm256,
+                        hpke::kdf::HkdfSha384,
+                        hpke::kem::DhP384HkdfSha384,
+                    >(pk_r, info, exporter_context, exporter_length, &mut *rng)?
+                }
+                _ => return Err(CryptoError::UnsupportedKem),
+            };
+
+        debug_assert_eq!(export.len(), exporter_length);
+
+        Ok((kem_output, export.into()))
     }
 
     fn hpke_setup_receiver_and_export(
@@ -455,30 +437,178 @@ impl OpenMlsCrypto for RustCrypto {
         info: &[u8],
         exporter_context: &[u8],
         exporter_length: usize,
-    ) -> Result<Vec<u8>, CryptoError> {
-        let context = hpke_from_config(config)
-            .setup_receiver(enc, &sk_r.into(), info, None, None, None)
-            .map_err(|_| CryptoError::ReceiverSetupError)?;
-        let exported_secret = context
-            .export(exporter_context, exporter_length)
-            .map_err(|_| CryptoError::ExporterError)?;
-        Ok(exported_secret)
+    ) -> Result<ExporterSecret, openmls_traits::types::CryptoError> {
+        let export =
+            match config {
+                HpkeConfig(HpkeKemType::DhKem25519, HpkeKdfType::HkdfSha256, HpkeAeadType::AesGcm128) => {
+                    hpke_core::hpke_export_rx::<
+                        hpke::aead::AesGcm128,
+                        hpke::kdf::HkdfSha256,
+                        hpke::kem::X25519HkdfSha256,
+                    >(enc, sk_r, info, exporter_context, exporter_length)?
+                }
+                HpkeConfig(HpkeKemType::DhKem25519, HpkeKdfType::HkdfSha256, HpkeAeadType::ChaCha20Poly1305) => {
+                    hpke_core::hpke_export_rx::<
+                        hpke::aead::ChaCha20Poly1305,
+                        hpke::kdf::HkdfSha256,
+                        hpke::kem::X25519HkdfSha256,
+                    >(enc, sk_r, info, exporter_context, exporter_length)?
+                }
+                HpkeConfig(HpkeKemType::DhKemP256, HpkeKdfType::HkdfSha256, HpkeAeadType::AesGcm128) => {
+                    hpke_core::hpke_export_rx::<
+                        hpke::aead::AesGcm128,
+                        hpke::kdf::HkdfSha256,
+                        hpke::kem::DhP256HkdfSha256,
+                    >(enc, sk_r, info, exporter_context, exporter_length)?
+                }
+                HpkeConfig(HpkeKemType::DhKemP384, HpkeKdfType::HkdfSha384, HpkeAeadType::AesGcm256) => {
+                    hpke_core::hpke_export_rx::<
+                        hpke::aead::AesGcm256,
+                        hpke::kdf::HkdfSha384,
+                        hpke::kem::DhP384HkdfSha384,
+                    >(enc, sk_r, info, exporter_context, exporter_length)?
+                }
+                _ => return Err(CryptoError::UnsupportedKem),
+            };
+
+        debug_assert_eq!(export.len(), exporter_length);
+
+        Ok(export.into())
     }
 
-    fn derive_hpke_keypair(&self, config: HpkeConfig, ikm: &[u8]) -> types::HpkeKeyPair {
-        let kp = hpke_from_config(config).derive_key_pair(ikm).unwrap().into_keys();
-        HpkeKeyPair {
-            private: kp.0.as_slice().into(),
-            public: kp.1.as_slice().into(),
+    fn derive_hpke_keypair(&self, config: HpkeConfig, ikm: &[u8]) -> Result<types::HpkeKeyPair, CryptoError> {
+        match config.0 {
+            HpkeKemType::DhKemP256 => hpke_core::hpke_derive_keypair::<hpke::kem::DhP256HkdfSha256>(ikm),
+            HpkeKemType::DhKemP384 => hpke_core::hpke_derive_keypair::<hpke::kem::DhP384HkdfSha384>(ikm),
+            HpkeKemType::DhKem25519 => hpke_core::hpke_derive_keypair::<hpke::kem::X25519HkdfSha256>(ikm),
+            _ => Err(CryptoError::UnsupportedKem),
         }
     }
 }
 
-fn hpke_from_config(config: HpkeConfig) -> Hpke<HpkeRustCrypto> {
-    Hpke::<HpkeRustCrypto>::new(
-        hpke::Mode::Base,
-        kem_mode(config.0),
-        kdf_mode(config.1),
-        aead_mode(config.2),
-    )
+mod hpke_core {
+    use openmls_traits::types::{CryptoError, HpkeCiphertext, HpkeKeyPair};
+
+    pub(crate) fn hpke_open<Aead: hpke::aead::Aead, Kdf: hpke::kdf::Kdf, Kem: hpke::Kem>(
+        private_key: &[u8],
+        kem_output: &[u8],
+        info: &[u8],
+        aad: &[u8],
+        ciphertext: &[u8],
+    ) -> Result<Vec<u8>, CryptoError> {
+        use hpke::Deserializable as _;
+        let encapped_key = Kem::EncappedKey::from_bytes(kem_output).map_err(|_| CryptoError::HpkeDecryptionError)?;
+        let key = Kem::PrivateKey::from_bytes(private_key).map_err(|_| CryptoError::HpkeDecryptionError)?;
+        let plaintext =
+            hpke::single_shot_open::<Aead, Kdf, Kem>(&hpke::OpModeR::Base, &key, &encapped_key, info, ciphertext, aad)
+                .map_err(|_| CryptoError::HpkeDecryptionError)?;
+
+        Ok(plaintext)
+    }
+
+    pub(crate) fn hpke_seal<Aead: hpke::aead::Aead, Kdf: hpke::kdf::Kdf, Kem: hpke::Kem>(
+        public_key: &[u8],
+        info: &[u8],
+        aad: &[u8],
+        plaintext: &[u8],
+        csprng: &mut impl rand_core::CryptoRngCore,
+    ) -> Result<HpkeCiphertext, CryptoError> {
+        use hpke::{Deserializable as _, Serializable as _};
+        let key = Kem::PublicKey::from_bytes(public_key).map_err(|_| CryptoError::HpkeEncryptionError)?;
+        let (encapped, ciphertext) =
+            hpke::single_shot_seal::<Aead, Kdf, Kem, _>(&hpke::OpModeS::Base, &key, info, plaintext, aad, csprng)
+                .map_err(|_| CryptoError::HpkeEncryptionError)?;
+
+        Ok(HpkeCiphertext {
+            kem_output: encapped.to_bytes().to_vec().into(),
+            ciphertext: ciphertext.into(),
+        })
+    }
+
+    #[allow(dead_code)]
+    pub(crate) fn hpke_gen_keypair<Kem: hpke::Kem>(
+        csprng: &mut impl rand_core::CryptoRngCore,
+    ) -> Result<HpkeKeyPair, CryptoError> {
+        use hpke::Serializable as _;
+        let (sk, pk) = Kem::gen_keypair(csprng);
+        let (private, public) = (sk.to_bytes().to_vec().into(), pk.to_bytes().to_vec());
+
+        Ok(HpkeKeyPair { private, public })
+    }
+
+    pub(crate) fn hpke_derive_keypair<Kem: hpke::Kem>(ikm: &[u8]) -> Result<HpkeKeyPair, CryptoError> {
+        use hpke::Serializable as _;
+        let (sk, pk) = Kem::derive_keypair(ikm);
+        let (private, public) = (sk.to_bytes().to_vec().into(), pk.to_bytes().to_vec());
+
+        Ok(HpkeKeyPair { private, public })
+    }
+
+    pub(crate) fn hpke_export_rx<Aead: hpke::aead::Aead, Kdf: hpke::kdf::Kdf, Kem: hpke::Kem>(
+        encapped_key: &[u8],
+        rx_private_key: &[u8],
+        info: &[u8],
+        export_info: &[u8],
+        export_len: usize,
+    ) -> Result<Vec<u8>, CryptoError> {
+        use hpke::Deserializable as _;
+        let key = Kem::PrivateKey::from_bytes(rx_private_key).map_err(|_| CryptoError::ReceiverSetupError)?;
+        let encapped_key = Kem::EncappedKey::from_bytes(encapped_key).map_err(|_| CryptoError::ReceiverSetupError)?;
+        let ctx = hpke::setup_receiver::<Aead, Kdf, Kem>(&hpke::OpModeR::Base, &key, &encapped_key, info)
+            .map_err(|_| CryptoError::ReceiverSetupError)?;
+
+        let mut export = vec![0u8; export_len];
+
+        ctx.export(export_info, &mut export)
+            .map_err(|_| CryptoError::ExporterError)?;
+
+        Ok(export)
+    }
+
+    pub(crate) fn hpke_export_tx<Aead: hpke::aead::Aead, Kdf: hpke::kdf::Kdf, Kem: hpke::Kem>(
+        tx_public_key: &[u8],
+        info: &[u8],
+        export_info: &[u8],
+        export_len: usize,
+        csprng: &mut impl rand_core::CryptoRngCore,
+    ) -> Result<(Vec<u8>, Vec<u8>), CryptoError> {
+        use hpke::{Deserializable as _, Serializable as _};
+        let key = Kem::PublicKey::from_bytes(tx_public_key).map_err(|_| CryptoError::SenderSetupError)?;
+        let (kem_output, ctx) = hpke::setup_sender::<Aead, Kdf, Kem, _>(&hpke::OpModeS::Base, &key, info, csprng)
+            .map_err(|_| CryptoError::SenderSetupError)?;
+
+        let mut export = vec![0u8; export_len];
+
+        ctx.export(export_info, &mut export)
+            .map_err(|_| CryptoError::ExporterError)?;
+
+        Ok((kem_output.to_bytes().to_vec(), export))
+    }
+}
+
+impl OpenMlsRand for RustCrypto {
+    type Error = MlsProviderError;
+
+    type RandImpl = rand_chacha::ChaCha20Rng;
+    type BorrowTarget<'a> = std::sync::RwLockWriteGuard<'a, Self::RandImpl>;
+
+    fn borrow_rand(&self) -> Result<Self::BorrowTarget<'_>, Self::Error> {
+        self.rng.write().map_err(|_| MlsProviderError::RngLockPoison)
+    }
+
+    fn random_array<const N: usize>(&self) -> Result<[u8; N], Self::Error> {
+        let mut rng = self.borrow_rand()?;
+        let mut out = [0u8; N];
+        rng.try_fill_bytes(&mut out)
+            .map_err(|_| MlsProviderError::UnsufficientEntropy)?;
+        Ok(out)
+    }
+
+    fn random_vec(&self, len: usize) -> Result<Vec<u8>, Self::Error> {
+        let mut rng = self.borrow_rand()?;
+        let mut out = vec![0u8; len];
+        rng.try_fill_bytes(&mut out)
+            .map_err(|_| MlsProviderError::UnsufficientEntropy)?;
+        Ok(out)
+    }
 }
